requirements.txt
PyQt6==6.5.0
mysql-connector-python==8.3.0
matplotlib==3.8.2